[refactor] seperate captcha
